feat: implement DetectNeedsStep and InsertImportsStep for import detection and insertion
